pyproject.toml
[project]
name = "et-market-oracle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt<4.0.0",
    "celery>=5.6.2",
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "feedparser>=6.0.12",
    "langchain>=1.2.13",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "passlib>=1.7.4",
    "pinecone>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
